User has same id as associated actor.
Idea is that user schema simply hold metadata for authentication and
authorization. The actor object is a RDF Source.

The reason for not embedding the Actor description in the User schema is to be
able to support non-User actors in the future (e.g. a group). This is probably
thinking ahead to far...