CPub.Web.Authentication.Strategy.Mastodon: Rename from `Fediverse` and `Pleroma` and use mnesia
The name `Mastodon` is more applicable as it is the OAuth protocol
defined by Mastodon.

Better not call it `Pleroma` or `Fediverse` as Pleroma developers would
probably have agreed on an already standardized way of doing this (e.g.
IndieAuth or OAuth 2.1) and in the future there might be an agreed upon
way of doing OAuth in the Fediverse.